requirements.txt
bs4==0.0.1
pymysql==0.10.1
requests==2.25.1
pycryptodome==3.12.0
beautifulsoup4~=4.11.1
urllib3~=1.26.9